+++ engine/src/main/kotlin/com/oregontrail/engine/GameRequests.kt
@@ -67,3 +67,7 @@ fun Game.clearQuickSaveRequest() {
 fun Game.clearQuickLoadRequest() {
         requestedQuickLoad = false
     }
+
+fun Game.clearJournalShareRequest() {
+    requestedJournalShare = false
+}